WIP.  Applied feedback.
